Move files related to EPLB (#7580)
